requirements.txt
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
et-xmlfile==1.0.1
google-api-core==1.22.2
google-auth==1.21.1
google-cloud-vision==1.0.0
googleapis-common-protos==1.52.0
grpcio==1.31.0
idna==2.10
jdcal==1.4.1
numpy==1.19.2
opencv-python==4.4.0.42
pandas==1.1.2
pdf2image==1.14.0
Pillow==7.2.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
rsa==4.6
scipy==1.5.2
six==1.15.0
urllib3==1.25.10
xlrd==1.2.0
